+++ packages/web/src/state/collection.ts
@@ -1,6 +1,6 @@
 import type { Description, Layout, LayoutName } from '@aeriajs/types'
-import { computed, reactive } from 'vue'
-import { useStore, type StoreState, type UnRef, type GlobalStateManager } from '@aeria-ui/state-management'
+import { computed, reactive, type ComputedRef } from 'vue'
+import { useStore, type StoreState, type UnwrapGetters, type GlobalStateManager } from '@aeria-ui/state-management'
 import { deepClone, deepMerge, isReference, getReferenceProperty } from '@aeriajs/common'
 import { PAGINATION_PER_PAGE_DEFAULT } from '../constants.js'
 import { deepDiff } from './deepDiff.js'
@@ -18,7 +18,7 @@ import {
 
 export type CollectionStoreState<TItem extends CollectionStoreItem = any> =
   ReturnType<typeof internalCreateCollectionStore<TItem>>['state']
-  & UnRef<ReturnType<ReturnType<typeof internalCreateCollectionStore>['getters']>>
+  & UnwrapGetters<ReturnType<ReturnType<typeof internalCreateCollectionStore>['getters']>>
 
 export type CollectionStore<TItem extends CollectionStoreItem = any> = CollectionStoreState<TItem> & {
   $id: string
@@ -27,7 +27,7 @@ export type CollectionStore<TItem extends CollectionStoreItem = any> = Collectio
 }
 
 export type CollectionStoreItem = Record<string, any> & {
-  _id?: string
+  _id?: any
 }
 
 const internalCreateCollectionStore = <TItem extends CollectionStoreItem>() => {
@@ -299,7 +299,7 @@ const internalCreateCollectionStore = <TItem extends CollectionStoreItem>() => {
 export const createCollectionStore = <TItem extends CollectionStoreItem>() => <
   TStoreId extends string,
   TStoreState extends StoreState = any,
-  TStoreGetters extends Record<string, ()=> any> = {},
+  TStoreGetters extends Record<string, (()=> any) | ComputedRef<any>> = {},
   TStoreActions extends Record<string, (...args: any[])=> any>={},
 >(
   newer: {
@@ -335,8 +335,8 @@ export const createCollectionStore = <TItem extends CollectionStoreItem>() => <
 
   return {
     $id: newer.$id,
-    state: state,
-    getters: newer.getters?.(state, actions as any) as TStoreGetters,
+    state: state as TStoreState & typeof initial['state'],
+    getters: newer.getters?.(state, actions as any) as TStoreGetters & ReturnType<typeof initial['getters']>,
     actions: actions as TStoreActions extends {}
       ? typeof actions & TStoreActions
       : never,