+++ packages/aeria-icons/CHANGELOG.md
@@ -1,5 +1,11 @@
 # aeria-icons
 
+## 0.0.2
+
+### Patch Changes
+
+- Several minor fixes
+
 ## 0.0.1
 
 ### Patch Changes